// src/app/interfaces/order.ts
export interface Order {
    id: number;
    customerId: number;
    dishId: number;
    finished: boolean;
}

export interface OrdersInterface {
    dan: string;
    restoranid: number;
    restoran: string;
    userrestoranuid: number;
    jelo: string;
    soup: boolean;
    Salad: boolean;
    bread: boolean;
    narucitelj: string;
    firma: string;
    firmaid: number;
    naruciteljid: number;
    finished: boolean;
    delivered: boolean;
  }